Major Home mode refactor to do menus better
I redid a ton of the Home mode to implement submenu support and make it
look cleaner. You can now select LED modes as well by navigating through
them and selecting what you want to change it to.

Navigation is:
  - Handshake 1: enter/exit (brings up the menu in Home mode)
  - Handshake 2: up/left
  - Handshake 3: down/right
  - Handshake 4: select

Special handling:
  In the "Display Handle" mode you press handshake 4 to edit your
  handle. Due to the limited supply of inputs I got creative. Here are
  the additional usages:
    - press handshakes 2 and 3 simultaneously for backspace while in
      text edit mode
    - press handshake 1 to save and exit the edit mode (there is no
      cancel because my creativity with limited inputs ran out at the
      time)

NOTE: with regard to the inputs there is a hold-to-repeat feature I
built so that you can press and hold an input to have it keep repeating
the same operation - useful for scrolling mostly