Build/flash docs and one code change
Changed the mad hatter mode so that hitting handshake 1 just exits back
to Home mode. Maybe we should do that with other modes too instead of
having to do the long press to force them to exit. For now I just have
the behavior for mad hatter now and for the about screen.

Added docs explaining how to package up firmware binaries and how to
flash them without needing to install things like PlatformIO. I also am
adding firmware images for the firmware build as of this commit.