Add offsets for UI rendering
We needed a way to add at least a Y offset because the displays we have
aren't aligned super well to the PCB cutout. This adds that support and
sets a 4px Y offset so that things _look_ vertically centered.

While I was in there I also implemented a feature I had forgotten about
with the text entry component to make the cursor remain on screen when
the text is too long for the screen width (ie. render the entered text
with a negative X offset so that the text and cursor can still be seen).

I'm also removing the `builds` directory and instead switching to tagged
releases with proper GitHub releases that contain the firmware archives.